+++ lib/radix_encoding.rb
@@ -1,8 +1,13 @@
 # frozen_string_literal: true
 
+require "radix_encoding/encoding"
 require "radix_encoding/version"
 
 module RadixEncoding
-  class Error < StandardError; end
-  # Your code goes here...
+  # Standard Base32 encoding as defined in RFC 4648.
+  BASE32 = Encoding.new(
+    alphabet: "ABCDEFGHIJKLMNOPQRSTUVWXYZ234567",
+    radix: 32,
+    padding: "=",
+  )
 end